requirements.txt
Cython==0.29.36
einops==0.7.0
h5py==3.8.0
librosa==0.10.1
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.1
phonemizer==3.2.1
PyYAML==6.0.1
scipy==1.11.4
torch==2.1.1
torchaudio==2.1.1
tqdm==4.66.1
Unidecode==1.3.6
wandb==0.15.2
typing_extensions==4.7.1
wandb==0.15.2
matplotlib==3.7.1
text-unidecode==1.3
Unidecode==1.3.6
phonemizer==3.2.1
librosa==0.10.1